% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmb_html.R
\name{margin_note}
\alias{margin_note}
\title{Function to add style to margin note}
\usage{
margin_note(text)
}
\arguments{
\item{text}{Text of margin note}
}
\description{
Function to add style to margin note
}
