% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmb_html.R
\name{msmb_html_book}
\alias{msmb_html_book}
\alias{msmb_html}
\title{Modern Statstics for Modern Biology HTML format}
\usage{
msmb_html_book(...)

msmb_html(..., margin_references = TRUE)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{bookdown::html_chapters()}}

\item{margin_references}{\code{logical}.  Determines whether to place 
citations in the margin, or collate them at the end of the document.}
}
\description{
Template for creating webpages in the style of Modern Statistics for
Modern Biology \url{http://www-huber.embl.de/msmb/}.
}
\details{
\code{msmb_html()} provides the HTML format used in 
Modern Statistics for Modern Biology:
  \url{http://www-huber.embl.de/msmb/}.

\code{msmb_html_book()} is used to create output split
into individual chapters for a book.  If you wish to create a complete
book this should be used in the output section of your YAML header
in the R Markdown document.
}
